# Evm stack improvements