This is a draft. Please do not review nor merge it.

This PR tries to improve Evm stack behavior by:

1. Moving throwing overflow/underflow exceptions to separate methods, that should make EvmStack methods much easier to inline for JIT
1. Treating data once as `Span<byte>` and once as 32 bytes word
1. Benchmarking it with a few new benchmarks

``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.17134.1069 (1803/April2018Update/Redstone4)
Intel Core i7-6700HQ CPU 2.60GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
Frequency=2531252 Hz, Resolution=395.0614 ns, Timer=TSC
.NET Core SDK=3.1.100
  [Host]     : .NET Core 3.0.1 (CoreCLR 4.700.19.51502, CoreFX 4.700.19.51609), X64 RyuJIT
  DefaultJob : .NET Core 3.0.1 (CoreCLR 4.700.19.51502, CoreFX 4.700.19.51609), X64 RyuJIT
```

#### Nethermind.Evm.Benchmark.EvmPopIntBenchmarks

|  Method |     Mean |   Error |  StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |---------:|--------:|--------:|------:|-------:|------:|------:|----------:|
| Current | 192.6 ns | 2.12 ns | 1.98 ns |  1.00 | 0.0305 |     - |     - |      96 B |
| After   | 230.7 ns | 3.59 ns | 3.35 ns |  1.00 | 0.0305 |     - |     - |      96 B |

#### Nethermind.Evm.Benchmark.EvmPushPopByteBenchmarks

|  Method |     Mean |    Error |   StdDev | Ratio | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |---------:|---------:|---------:|------:|------:|------:|------:|----------:|
| Current | 16.05 ns | 0.250 ns | 0.209 ns |  1.00 |     - |     - |     - |         - |
| After   | 13.66 ns | 0.278 ns | 0.330 ns |  1.00 |     - |     - |     - |         - |

#### Nethermind.Evm.Benchmark.EvmPushUInt256Benchmarks

|  Method |                value |     Mean |    Error |   StdDev | Ratio | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |--------------------- |---------:|---------:|---------:|------:|------:|------:|------:|----------:|
| Current |                    1 | 12.48 ns | 0.251 ns | 0.210 ns |  1.00 |     - |     - |     - |         - |
| After   |                    1 | 13.96 ns | 0.164 ns | 0.145 ns |  1.00 |     - |     - |     - |         - |
| Current | 11579(...)39935 [78] | 12.55 ns | 0.133 ns | 0.124 ns |  1.00 |     - |     - |     - |         - |
| After   | 11579(...)39935 [78] | 13.84 ns | 0.245 ns | 0.229 ns |  1.00 |     - |     - |     - |         - |