@Scooletz 2020-01-20
Dropping. Definitely not worth it! 

FYI @tkstanczak